refactor: extract shared mail seed context, split forward into mail_forward.go
Address review feedback:
- Extract resolveMailItemSeedContext shared by reply and forward seeds
- Move forward handler and thread-append logic to mail_forward.go
- Remove unused project variable suppressions in tests
- Replace threadEscapeHtml with shared escapeHtml from env
- Escape numeric HTML attributes defensively
